--- src/Login.js
@@ -1,39 +1,0 @@
-import React from "react"
-
-const Login = () => {
-    const firebase = require('firebase');
-    const firebaseui = require('firebaseui');
-    const ui = new firebaseui.auth.AuthUI(firebase.auth());
-    const uiConfig = {
-        callbacks: {
-          signInSuccessWithAuthResult: function(authResult, redirectUrl) {
-            // User successfully signed in.
-            // Return type determines whether we continue the redirect automatically
-            // or whether we leave that to developer to handle.
-            return true;
-          },
-          uiShown: function() {
-            // The widget is rendered.
-            // Hide the loader.
-            document.getElementById('loader').style.display = 'none';
-          }
-        },
-        // Will use popup for IDP Providers sign-in flow instead of the default, redirect.
-        signInFlow: 'popup',
-        signInSuccessUrl: '/',
-        signInOptions: [
-          firebase.auth.GoogleAuthProvider.PROVIDER_ID,
-          firebase.auth.EmailAuthProvider.PROVIDER_ID
-        ]
-      };
-      ui.start('#firebaseui-auth-container', uiConfig);
-
-    return (
-        <>
-        <div id="firebaseui-auth-container"></div>
-        <div id="loader">Loading...</div>
-        </>
-    )
-}
-
-export default Login
