+++ src/Perennial.js
@@ -1,7 +1,5 @@
 import React, {useState, useEffect} from "react"
 import AppViews from "../src/AppViews"
-import fire from "../src/config/fire"
-import Login from "../src/Login"
 
 const Perennial = (props) => {
     return (
